* quartz/ui_color.c, win32/ui_color.c: Cast u_int8_t to u_int32_t in   shifting bits by <<.   (https://github.com/arakiken/mlterm/issues/139) * ui_screen_manager.c, vt_term_manager.c: 0x1 -> 0x1u   (https://github.com/arakiken/mlterm/issues/139) * install.sh, main-cocoa.m: Move config files to mlterm.app/Contents/Resources/mlterm   instead of mlterm.app/mlterm.   (https://github.com/arakiken/mlterm/issues/141)

+++ uitoolkit/win32/ui_color.c
@@ -51,7 +51,7 @@ int ui_load_named_xcolor(ui_display_t *disp, ui_color_t *xcolor, const char *nam
 
 int ui_load_rgb_xcolor(ui_display_t *disp, ui_color_t *xcolor, u_int8_t red, u_int8_t green,
                        u_int8_t blue, u_int8_t alpha) {
-  xcolor->pixel = RGB(red, green, blue) | (alpha << 24);
+  xcolor->pixel = RGB(red, green, blue) | ((u_int32_t)alpha << 24);
 
   return 1;
 }
